requirements.txt
alabaster==0.7.12
aniso8601==9.0.1
antiorm==1.2.1
appdirs==1.4.4
asgiref==3.3.4
astor==0.8.1
asttokens==2.0.5
atomicwrites==1.4.0
attrs==20.3.0
automium==0.2.6
automium-web==0.1.1
awsebcli==3.19.4
Babel==2.9.1
bandit==1.7.0
beautifulsoup4==4.5.3
behave==1.2.5
botocore==1.20.47
cement==2.8.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.2.5
coverage==5.5
darglint==1.8.0
db==0.1.1
distlib==0.3.1
Django==3.2
docutils==0.17.1
eradicate==1.0
executing==0.6.0
filelock==3.0.12
flake8==3.9.1
flake8-bandit==2.1.2
flake8-broken-line==0.2.1
flake8-bugbear==19.8.0
flake8-commas==2.0.0
flake8-comprehensions==3.4.0
flake8-debugger==3.2.1
flake8-docstrings==1.6.0
flake8-eradicate==0.3.0
flake8-isort==3.0.1
flake8-polyfill==1.0.2
flake8-quotes==2.1.2
flake8-rst-docstrings==0.0.12
flake8-string-format==0.2.3
Flask==1.1.2
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
future==0.16.0
gitdb==4.0.7
GitPython==3.1.14
glob2==0.7
greenlet==1.0.0
heartrate==0.2.1
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
legacy==0.1.6
loguru==0.5.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.9
packit==0.28.0
parse==1.19.0
parse-type==0.5.2
pathspec==0.5.9
pbr==5.5.1
pep8-naming==0.9.1
pipenv==2020.11.15
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pycodestyle==2.7.0
pydocstyle==5.1.1
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==1.4.2
pyparsing==2.4.7
pypiwin32==223
pytest==6.2.3
pytest-cov==2.11.1
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
pyupgrade==2.13.0
pywin32==300
PyYAML==5.3.1
regex==2021.4.4
requests==2.25.1
restructuredtext-lint==1.3.2
selenium==3.141.0
semantic-version==2.8.5
six==1.15.0
smmap==4.0.0
snowballstemmer==2.1.0
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.4.12
sqlparse==0.4.1
stevedore==3.3.0
termcolor==1.1.0
testfixtures==6.17.1
tokenize-rt==4.1.0
toml==0.10.2
tox==3.23.0
tox-travis==0.12
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.25.11
virtualenv==20.4.3
virtualenv-clone==0.5.4
wcwidth==0.1.9
wemake-python-styleguide==0.14.1
Werkzeug==1.0.1
win32-setctime==1.0.3
